Add venv scripts to PATH in hook template
When invoking pre-commit from git hooks, if pre-commit was installed
from a virtual environment, activate the virtual environment by
prepending the script directory to PATH. This allows pre-commit to
invoke hooks installed into the same virtual environment it's running
from.

Rationale
---------

Some projects prefer dedicated tools to manage environments and
dependencies for Python-language hooks. pre-commit already supports this
via repository-local system hooks, but requires that the virtual
environment be activated in the calling shell.

In some situations, explicitly activating the environment is not
practical. Consider these two examples:

Example 1: pre-commit and [Nox]

  When running pre-commit in a Nox session, Python hooks can be
  installed into the environment associated with the Nox session,
  alongside pre-commit itself. This works fine when pre-commit is
  invoked by Nox: Nox activates the virtual environment for the
  pre-commit session, so the hooks are located on PATH. However, when
  pre-commit is invoked from git hooks, the virtual environment is not
  active, so pre-commit fails to locate the hooks.

Example 2: pre-commit and [Poetry]

  Another example is the case of adding pre-commit and the Python hooks
  as dependencies to a Poetry-managed project. When invoking pre-commit
  with Poetry (`poetry run pre-commit`), the environment is activated by
  Poetry. But when pre-commit is invoked from git hooks, the hooks are
  no longer on PATH.

[Nox]: https://nox.thea.codes
[Poetry]: https://python-poetry.org

Implementation Notes
--------------------

The hook template checks if the Python interpreter used when installing
pre-commit (`INSTALL_PYTHON`) is located in a virtual environment, by
checking for the file `pyvenv.cfg` two directories above the executable.
If this is the case, the directory containing the Python interpreter is
prepended to PATH. In a typical virtualenv, this is `bin` on UNIX-like
systems, and `Scripts` on Windows.

To keep the script short, the environment is modified in place, rather
than copied and passed to `os.execvpe` or `subprocess.Popen`.